refactor(public-form): shifted handleFormAuthRedirect from auth to public forms
